Reworked implementation of ReturnObject so that the constructor requires two arguments. One object type to be wrapped and one ErrorMessage enum type.
Responsibility for deciding which error message is appropriate
will be with the calling code, and must be determined before
the ReturnObject is instantiated.

I have included some code to automatically set the ErrorMessage
instance variable to INVALID_ARGUMENT if the object to be wrapped
is null, although I expect this to be redundant in the final
implementation.